refactor: move style variables to one file
