--- src/style/colorVariables.js
@@ -1,23 +1,0 @@
-export const COLOR_BRAND_PRIMARY = "#0aab2a";
-export const COLOR_BRAND_SECONDARY = "#002e6e";
-
-export const COLOR_RED = "#E4251A";
-export const COLOR_BLACK = "#000";
-export const COLOR_GREEN = "#0aab2a";
-export const COLOR_WHITE = "#fff";
-export const COLOR_GREY = "#444";
-
-export const COLOR_HIGHLIGHT = COLOR_BRAND_PRIMARY;
-export const COLOR_SUCCESS = "#0aab2a";
-export const COLOR_PRIMARY = "#fff";
-export const COLOR_DANGER = COLOR_RED;
-export const COLOR_WARNING = "#FAC60F";
-export const COLOR_INFO = "blue";
-
-export const DARK_ON_LIGHT_CONTRAST_ENHANCEMENT_RATIO = 2;
-
-// Theme Configuration
-export const COLOR_SIDEBAR_BACKGROUND = COLOR_BLACK;
-export const COLOR_CONTENT_BACKGROUND = COLOR_WHITE;
-
-export const BORDER_RADIUS_BASE = "3px";
