--- src/style/sizeVariables.js
@@ -1,7 +1,0 @@
-export const ICON_VIEWBOX_SIZE = 24;
-
-// z-index mappings
-export const ZINDEX_TOOLTIP = "1070";
-
-export const PADDING_BASE = "8px";
-
